+++ CHANGELOG.rst
@@ -2,10 +2,10 @@
 python-stdnet 0.3.0
 ========================================
  * Overall code refactoring.
- * Added ListField with Redis implementation.
+ * Added ListField and OrderedSetField with Redis implementation.
  * StdModel raise AttributError when method/attribute not available. Previously it returned None.
  * StdModel raise ModelNotRegistered when trying to save an instance of a non-registered model.
- * 13 tests.
+ * 17 tests.
 
 python-stdnet 0.2.2 - 2010 July 7
 ========================================